update snomed mapping
The updated mapping should now include all of the structures listed in
totalsegmentator/map_to_binary.py.

To create this map, the content of the original SNOMED mapping CSV was merged
with the exhaustive list of the labels from map_to_binary.py, and the mapping
for the missing structures was populated.