Replace set of if/elif with dictionary.
The set of if/elif mimics the switch construct which is not explicitly
available in Python. A more pythonic approach is to use a dictionary
and try-except.

This commit also addresses the issue of providing an invalid
class_type resulting in a NameError. In the original code the masks
name would not be set and the code tries to access it after the
if/elif section.